# Introduce <bit> shim-header