--- a/include/seqan3/alignment/matrix/edit_distance_score_matrix_full.hpp
+++ b/include/seqan3/alignment/matrix/edit_distance_score_matrix_full.hpp
@@ -109,7 +109,7 @@ public:
     static size_t max_rows(word_type const score_mask, unsigned const last_block,
                            score_type const score, score_type const max_errors) noexcept
     {
-        size_t const offset = score_mask == 0u ? 0u : std::bit_width(score_mask);
+        size_t const offset = std::bit_width(score_mask);
         size_t const active_row = word_size * last_block + offset;
         return active_row + (score <= max_errors);
     }