--- a/include/seqan3/core/bit_manipulation.hpp
+++ b/include/seqan3/core/bit_manipulation.hpp
@@ -189,7 +189,7 @@ SEQAN3_DEPRECATED_310 constexpr uint8_t count_trailing_zeros(unsigned_t const n)
  * Constant.
  */
 template <std::unsigned_integral unsigned_t>
-constexpr uint8_t most_significant_bit_set(unsigned_t const n) noexcept
+SEQAN3_DEPRECATED_310 constexpr uint8_t most_significant_bit_set(unsigned_t const n) noexcept
 {
     assert(n > 0); // n == 0 might have undefined behaviour
     return std::bit_width(n) - 1;