--- a/include/seqan3/core/bit_manipulation.hpp
+++ b/include/seqan3/core/bit_manipulation.hpp
@@ -125,7 +125,7 @@ SEQAN3_DEPRECATED_310 constexpr uint8_t popcount(unsigned_t const n) noexcept
 * Constant.
 */
 template <std::unsigned_integral unsigned_t>
-constexpr uint8_t count_leading_zeros(unsigned_t const n) noexcept
+SEQAN3_DEPRECATED_310 constexpr uint8_t count_leading_zeros(unsigned_t const n) noexcept
 {
     assert(n > 0); // n == 0 might have undefined behaviour
     return std::countl_zero(n);