--- a/include/seqan3/range/container/dynamic_bitset.hpp
+++ b/include/seqan3/range/container/dynamic_bitset.hpp
@@ -12,12 +12,14 @@
 
 #pragma once
 
-#include <seqan3/core/bit_manipulation.hpp>
+#include <seqan3/std/bit>
+
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
 #include <seqan3/range/views/interleave.hpp>
 #include <seqan3/range/views/repeat_n.hpp>
 #include <seqan3/range/views/to.hpp>
+#include <seqan3/utility/detail/integer_traits.hpp>
 
 namespace seqan3
 {
@@ -214,7 +216,7 @@ public:
         if (std::popcount(value >> 58u) != 0u)
             throw std::invalid_argument{"The dynamic_bitset can be at most 58 long."};
         data.bits |= value;
-        data.size |= value ? detail::most_significant_bit_set(value) + 1 : 0u;
+        data.size |= value ? std::bit_width(value) : 0u;
     }
 
     /*!\brief Construct from two iterators.