--- a/include/seqan3/search/dream_index/interleaved_bloom_filter.hpp
+++ b/include/seqan3/search/dream_index/interleaved_bloom_filter.hpp
@@ -227,7 +227,7 @@ public:
         if (bin_size_ == 0)
             throw std::logic_error{"The size of a bin must be > 0."};
 
-        hash_shift = detail::count_leading_zeros(bin_size_);
+        hash_shift = std::countl_zero(bin_size_);
         bin_words = (bins + 63) >> 6; // = ceil(bins/64)
         technical_bins  = bin_words << 6; // = bin_words * 64
         data = sdsl::bit_vector(technical_bins * bin_size_);