--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -9,6 +9,7 @@
 /*!\file
  * \brief Provides the C++20 \<bit\> header if it is not already available.
  * \author Enrico Seiler <enrico.seiler AT fu-berlin.de>
+ * \author Marcel Ehrhardt <marcel.ehrhardt AT fu-berlin.de>
  */
 
 #pragma once
@@ -49,3 +50,32 @@ enum class endian
 } //namespace std
 
 #endif // __cpp_lib_endian
+
+#ifndef __cpp_lib_int_pow2
+
+namespace std
+{
+
+// // bit_­cast
+// template<class To, class From> constexpr To bit_cast(const From& from) noexcept;
+//
+// // integral powers of 2
+// template<class T> constexpr bool has_single_bit(T x) noexcept;
+// template<class T> constexpr T bit_ceil(T x) noexcept;
+// template<class T> constexpr T bit_floor(T x) noexcept;
+// template<class T> constexpr T bit_width(T x) noexcept;
+//
+// // rotating
+// template<class T> [[nodiscard]] constexpr T rotl(T x, int s) noexcept;
+// template<class T> [[nodiscard]] constexpr T rotr(T x, int s) noexcept;
+//
+// // counting
+// template<class T> constexpr int countl_zero(T x) noexcept;
+// template<class T> constexpr int countl_one(T x) noexcept;
+// template<class T> constexpr int countr_zero(T x) noexcept;
+// template<class T> constexpr int countr_one(T x) noexcept;
+// template<class T> constexpr int popcount(T x) noexcept;
+
+} // namespace std
+
+#endif // __cpp_lib_int_pow2