--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -17,6 +17,7 @@
 #if __has_include(<bit>)
 #include <bit>
 #endif // __has_include(<bit>)
+#include <cassert>
 #include <climits>
 #include <type_traits>
 
@@ -111,7 +112,18 @@ template<class T> constexpr int countl_zero(T x) noexcept
 }
 
 // template<class T> constexpr int countl_one(T x) noexcept;
-// template<class T> constexpr int countr_zero(T x) noexcept;
+template<class T> constexpr int countr_zero(T x) noexcept
+{
+    static_assert(std::is_unsigned_v<T>, "Input should be unsigned.");
+    assert(x > 0); // x == 0 might have undefined behaviour
+
+    if constexpr (sizeof(T) == sizeof(unsigned long long))
+        return __builtin_ctzll(x);
+    else if constexpr (sizeof(T) == sizeof(unsigned long))
+        return __builtin_ctzl(x);
+    else
+        return __builtin_ctz(x);
+}
 // template<class T> constexpr int countr_one(T x) noexcept;
 // template<class T> constexpr int popcount(T x) noexcept;