--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -54,16 +54,16 @@ enum class endian
 
 #endif // __cpp_lib_endian
 
-#ifndef __cpp_lib_int_pow2
-
-namespace std
-{
-
-namespace detail
+namespace std::detail
 {
 template <typename type_t>
 constexpr auto sizeof_bits = CHAR_BIT * sizeof(type_t);
 } // namespace std::detail
+
+#ifndef __cpp_lib_int_pow2
+
+namespace std
+{
 
 // forward declare
 template<class T> constexpr int countl_zero(T x) noexcept;
@@ -100,6 +100,15 @@ template<class T> constexpr T bit_width(T x) noexcept
 
     return detail::sizeof_bits<T> - countl_zero(x);
 }
+
+} // namespace std
+
+#endif // __cpp_lib_int_pow2
+
+#ifndef __cpp_lib_bitops
+
+namespace std
+{
 //
 // // rotating
 // template<class T> [[nodiscard]] constexpr T rotl(T x, int s) noexcept;
@@ -145,4 +154,4 @@ template<class T> constexpr int popcount(T x) noexcept
 
 } // namespace std
 
-#endif // __cpp_lib_int_pow2
+#endif // __cpp_lib_bitops