--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -109,28 +109,26 @@ template<class T> constexpr T bit_width(T x) noexcept
 template<class T> constexpr int countl_zero(T x) noexcept
 {
     static_assert(std::is_unsigned_v<T>, "Input should be unsigned.");
-    assert(x > 0); // n == 0 might have undefined behaviour
 
     if constexpr (sizeof(T) == sizeof(unsigned long long))
-        return __builtin_clzll(x);
+        return x == 0u ? detail::sizeof_bits<T> : __builtin_clzll(x);
     else if constexpr (sizeof(T) == sizeof(unsigned long))
-        return __builtin_clzl(x);
+        return x == 0u ? detail::sizeof_bits<T> : __builtin_clzl(x);
     else
-        return __builtin_clz(x) + detail::sizeof_bits<T> - detail::sizeof_bits<unsigned int>;
+        return detail::sizeof_bits<T> + (x == 0u ? 0u : __builtin_clz(x) - detail::sizeof_bits<unsigned int>);
 }
 
 // template<class T> constexpr int countl_one(T x) noexcept;
 template<class T> constexpr int countr_zero(T x) noexcept
 {
     static_assert(std::is_unsigned_v<T>, "Input should be unsigned.");
-    assert(x > 0); // x == 0 might have undefined behaviour
 
     if constexpr (sizeof(T) == sizeof(unsigned long long))
-        return __builtin_ctzll(x);
+        return x == 0u ? detail::sizeof_bits<T> : __builtin_ctzll(x);
     else if constexpr (sizeof(T) == sizeof(unsigned long))
-        return __builtin_ctzl(x);
+        return x == 0u ? detail::sizeof_bits<T> : __builtin_ctzl(x);
     else
-        return __builtin_ctz(x);
+        return x == 0u ? detail::sizeof_bits<T> : __builtin_ctz(x);
 }
 // template<class T> constexpr int countr_one(T x) noexcept;
 template<class T> constexpr int popcount(T x) noexcept