--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -28,7 +28,7 @@
 #   if defined(__cpp_lib_bitops)
 #       define SEQAN3_CPP_LIB_BITOPS 1
 #   elif defined(__GNUC__) && (__GNUC__ == 9) && __cplusplus > 201703L /*C++2a*/
-#       define SEQAN3_CPP_LIB_BITOPS 1
+// #       define SEQAN3_CPP_LIB_BITOPS 1
 #   endif
 #endif
 
@@ -84,6 +84,9 @@ constexpr auto sizeof_bits = CHAR_BIT * sizeof(type_t);
 namespace std
 {
 
+namespace
+{
+
 #ifndef SEQAN3_CPP_LIB_BITOPS
 // forward declare
 template<class T> constexpr int countl_zero(T x) noexcept;
@@ -137,6 +140,8 @@ template<class T> constexpr T bit_width(T x) noexcept
     return detail::sizeof_bits<T> - countl_zero(x);
 }
 
+} // anonymous namespace
+
 } // namespace std
 
 #endif // __cpp_lib_int_pow2
@@ -145,6 +150,10 @@ template<class T> constexpr T bit_width(T x) noexcept
 
 namespace std
 {
+
+namespace
+{
+
 //
 // // rotating
 // template<class T> [[nodiscard]] constexpr T rotl(T x, int s) noexcept;
@@ -202,6 +211,8 @@ template<class T> constexpr int popcount(T x) noexcept
     else
         return __builtin_popcount(x);
 }
+
+} // anonymous namespace
 
 } // namespace std