--- a/include/seqan3/std/bit
+++ b/include/seqan3/std/bit
@@ -17,7 +17,6 @@
 #if __has_include(<bit>)
 #include <bit>
 #endif // __has_include(<bit>)
-#include <cassert>
 #include <climits>
 #include <type_traits>