--- a/include/seqan3/utility/math.hpp
+++ b/include/seqan3/utility/math.hpp
@@ -12,9 +12,9 @@
 
 #pragma once
 
+#include <seqan3/std/bit>
 #include <cassert>
 #include <cmath>
-#include <seqan3/std/bit>
 #include <seqan3/std/concepts>
 #include <stdexcept>