… include/seqan3/utility/simd/detail/builtin_simd.hpp
… test/performance/core/bit_manipulation_benchmark.cpp
… test/snippet/core/detail/count_leading_zeros.cpp
… test/snippet/core/detail/count_trailing_zeros.cpp
… test/snippet/core/detail/most_significant_bit_set.cpp
… test/snippet/core/detail/popcount.cpp
… test/unit/core/bit_manipulation_test.cpp
… test/unit/range/container/aligned_allocator_test.cpp
… test/unit/range/container/dynamic_bitset_test.cpp
… test/unit/utility/math_test.cpp