Fixes Partially https://github.com/seqan/product_backlog/issues/199

gcc-9 comparison:

| Benchmark                                                  | gcc-9-before Time | gcc-9-after Time | After-before Time | (After-Before)/Before Time |
| ---------------------------------------------------------- | ----------------- | ---------------- | ----------------- | -------------------------- |
| _is\_power\_of\_two\_popcount\<unsigned>\_median_           | 8.610 ns          | 9.170 ns         | 0.560 ns          | **+6.50%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_median_      | 7.880 ns          | 8.290 ns         | 0.410 ns          | **+5.20%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_median_ | 9.070 ns          | 8.860 ns         | \-0.210 ns        | \-2.32%                    |
| _is\_power\_of\_two\_arithmetic\_median_                   | 6.950 ns          | 6.920 ns         | \-0.030 ns        | \-0.43%                    |
| **is\_power\_of\_two\_seqan3\_median**                     | 6.870 ns          | 6.960 ns         | 0.090 ns          | +1.31%                     |
| _is\_power\_of\_two\_std\_median_                          |                   | 6.950 ns         |                   |                            |
| **next\_power\_of\_two\_seqan3\_median**                   | 14.400 ns         | 14.100 ns        | \-0.300 ns        | \-2.08%                    |
| **next\_power\_of\_two\_std\_median**                      |                   | 14.000 ns        |                   |                            |

<details><summary>Click for complete table</summary>

| Benchmark                                                  | gcc-9-before Time | gcc-9-after Time | After-before Time | (After-Before)/Before Time |
| ---------------------------------------------------------- | ----------------- | ---------------- | ----------------- | -------------------------- |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 8.540 ns          | 9.180 ns         | 0.640 ns          | **+7.49%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 8.610 ns          | 9.160 ns         | 0.550 ns          | **+6.39%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 8.650 ns          | 9.170 ns         | 0.520 ns          | **+6.01%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_mean_             | 8.600 ns          | 9.170 ns         | 0.570 ns          | **+6.63%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_median_           | 8.610 ns          | 9.170 ns         | 0.560 ns          | **+6.50%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_stddev_           | 0.056 ns          | 0.008 ns         | \-0.048 ns        | **\-85.71%**               |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.860 ns          | 8.350 ns         | 0.490 ns          | **+6.23%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.880 ns          | 8.280 ns         | 0.400 ns          | **+5.08%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.950 ns          | 8.290 ns         | 0.340 ns          | +4.28%                     |
| _is\_power\_of\_two\_popcount\<unsigned long>\_mean_        | 7.900 ns          | 8.310 ns         | 0.410 ns          | **+5.19%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_median_      | 7.880 ns          | 8.290 ns         | 0.410 ns          | **+5.20%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_stddev_      | 0.046 ns          | 0.036 ns         | \-0.010 ns        | **\-21.74%**               |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 8.900 ns          | 8.840 ns         | \-0.060 ns        | \-0.67%                    |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 9.100 ns          | 8.860 ns         | \-0.240 ns        | \-2.64%                    |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 9.070 ns          | 8.940 ns         | \-0.130 ns        | \-1.43%                    |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_mean_   | 9.020 ns          | 8.880 ns         | \-0.140 ns        | \-1.55%                    |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_median_ | 9.070 ns          | 8.860 ns         | \-0.210 ns        | \-2.32%                    |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_stddev_ | 0.107 ns          | 0.054 ns         | \-0.053 ns        | **\-49.53%**               |
| _is\_power\_of\_two\_arithmetic_                           | 7.040 ns          | 6.910 ns         | \-0.130 ns        | \-1.85%                    |
| _is\_power\_of\_two\_arithmetic_                           | 6.950 ns          | 6.950 ns         | 0.000 ns          | +0.00%                     |
| _is\_power\_of\_two\_arithmetic_                           | 6.910 ns          | 6.920 ns         | 0.010 ns          | +0.14%                     |
| _is\_power\_of\_two\_arithmetic\_mean_                     | 6.970 ns          | 6.920 ns         | \-0.050 ns        | \-0.72%                    |
| _is\_power\_of\_two\_arithmetic\_median_                   | 6.950 ns          | 6.920 ns         | \-0.030 ns        | \-0.43%                    |
| _is\_power\_of\_two\_arithmetic\_stddev_                   | 0.067 ns          | 0.020 ns         | \-0.047 ns        | **\-70.15%**               |
| **is\_power\_of\_two\_seqan3**                             | 6.870 ns          | 6.940 ns         | 0.070 ns          | +1.02%                     |
| **is\_power\_of\_two\_seqan3**                             | 6.870 ns          | 6.960 ns         | 0.090 ns          | +1.31%                     |
| **is\_power\_of\_two\_seqan3**                             | 6.890 ns          | 6.970 ns         | 0.080 ns          | +1.16%                     |
| **is\_power\_of\_two\_seqan3\_mean**                       | 6.880 ns          | 6.960 ns         | 0.080 ns          | +1.16%                     |
| **is\_power\_of\_two\_seqan3\_median**                     | 6.870 ns          | 6.960 ns         | 0.090 ns          | +1.31%                     |
| **is\_power\_of\_two\_seqan3\_stddev**                     | 0.012 ns          | 0.016 ns         | 0.004 ns          | **+33.33%**                |
| _is\_power\_of\_two\_std_                                  |                   | 6.940 ns         |                   |                            |
| _is\_power\_of\_two\_std_                                  |                   | 6.990 ns         |                   |                            |
| _is\_power\_of\_two\_std_                                  |                   | 6.950 ns         |                   |                            |
| _is\_power\_of\_two\_std\_mean_                            |                   | 6.960 ns         |                   |                            |
| _is\_power\_of\_two\_std\_median_                          |                   | 6.950 ns         |                   |                            |
| _is\_power\_of\_two\_std\_stddev_                          |                   | 0.022 ns         |                   |                            |
| **next\_power\_of\_two\_seqan3**                           | 14.300 ns         | 14.100 ns        | \-0.200 ns        | \-1.40%                    |
| **next\_power\_of\_two\_seqan3**                           | 14.400 ns         | 14.100 ns        | \-0.300 ns        | \-2.08%                    |
| **next\_power\_of\_two\_seqan3**                           | 14.400 ns         | 14.100 ns        | \-0.300 ns        | \-2.08%                    |
| **next\_power\_of\_two\_seqan3\_mean**                     | 14.400 ns         | 14.100 ns        | \-0.300 ns        | \-2.08%                    |
| **next\_power\_of\_two\_seqan3\_median**                   | 14.400 ns         | 14.100 ns        | \-0.300 ns        | \-2.08%                    |
| **next\_power\_of\_two\_seqan3\_stddev**                   | 0.036 ns          | 0.012 ns         | \-0.024 ns        | **\-66.67%**               |
| **next\_power\_of\_two\_std**                              |                   | 13.900 ns        |                   |                            |
| **next\_power\_of\_two\_std**                              |                   | 14.000 ns        |                   |                            |
| **next\_power\_of\_two\_std**                              |                   | 14.000 ns        |                   |                            |
| **next\_power\_of\_two\_std\_mean**                        |                   | 14.000 ns        |                   |                            |
| **next\_power\_of\_two\_std\_median**                      |                   | 14.000 ns        |                   |                            |

</details>

gcc-10 comparison:

| Benchmark                                                  | gcc-10-before Time | gcc-10-after Time | After-before Time | (After-Before)/Before Time |
| ---------------------------------------------------------- | ------------------ | ----------------- | ----------------- | -------------------------- |
| _is\_power\_of\_two\_popcount\<unsigned>\_median_           | 7.710 ns           | 8.290 ns          | 0.580 ns          | **+7.52%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_median_      | 7.770 ns           | 8.410 ns          | 0.640 ns          | **+8.24%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_median_ | 8.060 ns           | 8.360 ns          | 0.300 ns          | +3.72%                     |
| _is\_power\_of\_two\_arithmetic\_median_                   | 6.190 ns           | 6.350 ns          | 0.160 ns          | +2.58%                     |
| **is\_power\_of\_two\_seqan3\_median**                     | 6.160 ns           | 8.360 ns          | 2.200 ns          | **+35.71%**                |
| **is\_power\_of\_two\_std\_median**                        | 8.13               | 8.380 ns          | 0.250 ns          | +3.08%                     |
| **next\_power\_of\_two\_seqan3\_median**                   | 13.700 ns          | 6.700 ns          | \-7.000 ns        | **\-51.09%**               |
| _next\_power\_of\_two\_std\_median_                        | 6.54               | 6.680 ns          | 0.140 ns          | +2.14%                     |

<details><summary>Click for complete table</summary>

| Benchmark                                                  | gcc-10-before Time | gcc-10-after Time | After-before Time | (After-Before)/Before Time |
| ---------------------------------------------------------- | ------------------ | ----------------- | ----------------- | -------------------------- |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 7.700 ns           | 8.290 ns          | 0.590 ns          | **+7.66%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 7.710 ns           | 8.290 ns          | 0.580 ns          | **+7.52%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>_                   | 7.740 ns           | 8.290 ns          | 0.550 ns          | **+7.11%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_mean_             | 7.720 ns           | 8.290 ns          | 0.570 ns          | **+7.38%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_median_           | 7.710 ns           | 8.290 ns          | 0.580 ns          | **+7.52%**                 |
| _is\_power\_of\_two\_popcount\<unsigned>\_stddev_           | 0.018 ns           | 0.002 ns          | \-0.016 ns        | **\-88.89%**               |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.710 ns           | 8.420 ns          | 0.710 ns          | **+9.21%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.770 ns           | 8.400 ns          | 0.630 ns          | **+8.11%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>_              | 7.920 ns           | 8.410 ns          | 0.490 ns          | **+6.19%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_mean_        | 7.800 ns           | 8.410 ns          | 0.610 ns          | **+7.82%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_median_      | 7.770 ns           | 8.410 ns          | 0.640 ns          | **+8.24%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long>\_stddev_      | 0.110 ns           | 0.007 ns          | \-0.103 ns        | **\-93.64%**               |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 7.960 ns           | 8.370 ns          | 0.410 ns          | **+5.15%**                 |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 8.060 ns           | 8.350 ns          | 0.290 ns          | +3.60%                     |
| _is\_power\_of\_two\_popcount\<unsigned long long>_         | 8.150 ns           | 8.360 ns          | 0.210 ns          | +2.58%                     |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_mean_   | 8.060 ns           | 8.360 ns          | 0.300 ns          | +3.72%                     |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_median_ | 8.060 ns           | 8.360 ns          | 0.300 ns          | +3.72%                     |
| _is\_power\_of\_two\_popcount\<unsigned long long>\_stddev_ | 0.094 ns           | 0.010 ns          | \-0.084 ns        | **\-89.36%**               |
| _is\_power\_of\_two\_arithmetic_                           | 6.170 ns           | 6.350 ns          | 0.180 ns          | +2.92%                     |
| _is\_power\_of\_two\_arithmetic_                           | 6.190 ns           | 6.360 ns          | 0.170 ns          | +2.75%                     |
| _is\_power\_of\_two\_arithmetic_                           | 6.200 ns           | 6.350 ns          | 0.150 ns          | +2.42%                     |
| _is\_power\_of\_two\_arithmetic\_mean_                     | 6.190 ns           | 6.350 ns          | 0.160 ns          | +2.58%                     |
| _is\_power\_of\_two\_arithmetic\_median_                   | 6.190 ns           | 6.350 ns          | 0.160 ns          | +2.58%                     |
| _is\_power\_of\_two\_arithmetic\_stddev_                   | 0.013 ns           | 0.008 ns          | \-0.005 ns        | **\-38.46%**               |
| **is\_power\_of\_two\_seqan3**                             | 6.160 ns           | 8.350 ns          | 2.190 ns          | **+35.55%**                |
| **is\_power\_of\_two\_seqan3**                             | 6.160 ns           | 8.380 ns          | 2.220 ns          | **+36.04%**                |
| **is\_power\_of\_two\_seqan3**                             | 6.180 ns           | 8.360 ns          | 2.180 ns          | **+35.28%**                |
| **is\_power\_of\_two\_seqan3\_mean**                       | 6.170 ns           | 8.370 ns          | 2.200 ns          | **+35.66%**                |
| **is\_power\_of\_two\_seqan3\_median**                     | 6.160 ns           | 8.360 ns          | 2.200 ns          | **+35.71%**                |
| **is\_power\_of\_two\_seqan3\_stddev**                     | 0.010 ns           | 0.016 ns          | 0.006 ns          | **+60.00%**                |
| **is\_power\_of\_two\_std**                                | 8.17               | 8.470 ns          | 0.300 ns          | +3.67%                     |
| **is\_power\_of\_two\_std**                                | 8.13               | 8.380 ns          | 0.250 ns          | +3.08%                     |
| **is\_power\_of\_two\_std**                                | 8.02               | 8.370 ns          | 0.350 ns          | +4.36%                     |
| **is\_power\_of\_two\_std\_mean**                          | 8.11               | 8.410 ns          | 0.300 ns          | +3.70%                     |
| **is\_power\_of\_two\_std\_median**                        | 8.13               | 8.380 ns          | 0.250 ns          | +3.08%                     |
| **is\_power\_of\_two\_std\_stddev**                        | 0.077              | 0.055 ns          | \-0.022 ns        | **\-28.57%**               |
| **next\_power\_of\_two\_seqan3**                           | 13.700 ns          | 6.700 ns          | \-7.000 ns        | **\-51.09%**               |
| **next\_power\_of\_two\_seqan3**                           | 13.700 ns          | 6.700 ns          | \-7.000 ns        | **\-51.09%**               |
| **next\_power\_of\_two\_seqan3**                           | 13.800 ns          | 6.720 ns          | \-7.080 ns        | **\-51.30%**               |
| **next\_power\_of\_two\_seqan3\_mean**                     | 13.700 ns          | 6.710 ns          | \-6.990 ns        | **\-51.02%**               |
| **next\_power\_of\_two\_seqan3\_median**                   | 13.700 ns          | 6.700 ns          | \-7.000 ns        | **\-51.09%**               |
| **next\_power\_of\_two\_seqan3\_stddev**                   | 0.050 ns           | 0.015 ns          | \-0.035 ns        | **\-70.00%**               |
| _next\_power\_of\_two\_std_                                | 6.44               | 6.650 ns          | 0.210 ns          | +3.26%                     |
| _next\_power\_of\_two\_std_                                | 6.54               | 6.680 ns          | 0.140 ns          | +2.14%                     |
| _next\_power\_of\_two\_std_                                | 6.66               | 6.680 ns          | 0.020 ns          | +0.30%                     |
| _next\_power\_of\_two\_std\_mean_                          | 6.55               | 6.670 ns          | 0.120 ns          | +1.83%                     |
| _next\_power\_of\_two\_std\_median_                        | 6.54               | 6.680 ns          | 0.140 ns          | +2.14%                     |
| _next\_power\_of\_two\_std\_stddev_                        | 0.107              | 0.018             | \-0.089 ns        | **\-83.18%**               |

</details>

Note:
* _cursive_ Benchmarks: no change in implementation
* **bold** Benchmarks: change in implementation
* **bold** Time Percentage changes: has significant percentage change

For some reason the popcount benchmark (independent of seqan3) did
increase the runtime. I'm not sure what happened here. This is
consistent between gcc-9 and gcc-10.

Looking at the number:
* gcc-9 has the same runtimes (same implementation used)
* gcc-10 -std=c++2a has a better next_power_of_two implementation (std::bit_ceil), we should use the same implementation in <= gcc-9.
* gcc-10 -std=c++2a seems to have a worser implementation of is_power_of_two (std::has_single_bit), I need to report that upstream.