@marehr 2020-12-09
> I reran the benchmark on my machine with `-O3 -march=native` and did not see any performance regressions.
> I also noted the effect that gcc10 `next_power_of_two` is around 2 times faster (11.3 ns-> 5.3ns for me).

You are right, the std implementation is faster for -march=native, without that it is a bit slower.