@eseiler 2020-12-11
I think you use deprecated stuff in the benchmarks? At least it was the case when I tried the benchmarks locally.
Also, the stuff in `std/bit` is not documented, I think we usually do a short description and a link to cppreference (like in `std::endian`).