@marehr 2020-12-11
> I think you use deprecated stuff in the benchmarks? At least it was the case when I tried the benchmarks locally.
> Also, the stuff in `std/bit` is not documented, I think we usually do a short description and a link to cppreference (like in `std::endian`).

Thank you, normally the std:: entities should be auto documented. We currently use `#ifndef` which will open up our shim functions for doxygen.