@marehr 2020-12-13
Finally! I got it working, grrr....